applications/4.gene_expression_prediction_based_on_multi_modal_data/requirement.model.txt
# Python 3.10.12
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
deepspeed==0.16.9
dill==0.3.8
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.121.1
ffmpy==1.0.0
filelock==3.19.1
fire==0.7.1
flash_attn==2.8.3
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llamafactory==0.9.3
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
ninja==1.13.0
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
peft==0.15.2
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyBigWig==0.3.25
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyfaidx==0.9.0.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruff==0.14.4
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.43.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sse-starlette==3.0.3
starlette==0.49.3
sympy==1.13.1
tensorboardX==2.6.4
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.51.0
triton==3.1.0
trl==0.9.6
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.22.3
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
